+++ Breeze.Api/Services/ExpenseServices.cs
@@ -17,14 +17,14 @@ namespace Breeze.Api.Services
             db = dbContext;
         }
 
-        public List<ExpenseResponse> GetExpenseByCategoryId(int CategoryId)
+        public List<ExpenseResponse> GetExpenseByCategoryId(string userId, int CategoryId)
         {
             return db.Expenses
-                .Where(expense => expense.Category.Id == CategoryId)
+                .Where(expense => expense.Category.Id == CategoryId && expense.User.Id.Equals(userId))
                 .Select(expense => new ExpenseResponse
                 {
                     Id = expense.Category.Id,
-                    UserEmail = expense.User.UserEmail,
+                    UserId = expense.User.UserId,
                     Name = expense.Name,
                     Date = expense.Date,
                     CategoryId = CategoryId,
@@ -33,12 +33,12 @@ namespace Breeze.Api.Services
                 .ToList();
         }
 
-        public int CreateExpense(ExpenseRequest newExpense)
+        public int CreateExpense(string userId, ExpenseRequest newExpense)
         {
             try
             {
                 var category = db.Categories.Find(newExpense.CategoryId);
-                var user = db.Users.Find(newExpense.UserEmail);
+                var user = db.Users.Find(userId);
                 if (category == null || user == null)
                 {
                     return -1;
@@ -63,12 +63,17 @@ namespace Breeze.Api.Services
             }
         }
 
-        public int UpdateExpense(ExpenseRequest existingExpense)
+        public int UpdateExpense(string userId, ExpenseRequest existingExpense)
         {
             var expense = db.Expenses.Find(existingExpense.Id);
+
             if (expense == null)
             {
                 return -1;
+            }
+            if (!expense.User.Id.Equals(userId))
+            {
+                return -2;
             }
             try
             {
@@ -87,12 +92,16 @@ namespace Breeze.Api.Services
             }
         }
 
-        public int DeleteExpense(int expenseId)
+        public int DeleteExpense(string userId, int expenseId)
         {
             var expense = db.Expenses.Find(expenseId);
             if (expense == null)
             {
                 return -1;
+            }
+            if (!expense.User.Id.Equals(userId))
+            {
+                return -2;
             }
             db.Expenses.Remove(expense);
             db.SaveChanges();
